+++ src/Codeception/Module/Symfony2.php
@@ -22,6 +22,7 @@ use Symfony\Component\HttpFoundation\Response;
  *
  * * app_path: 'app' - specify custom path to your app dir, where bootstrap cache and kernel interface is located.
  * * environment: 'local' - environment used for load kernel
+ * * debug: true - switch debug mode
 * 
  * ### Example (`functional.suite.yml`)
  *
@@ -31,6 +32,7 @@ use Symfony\Component\HttpFoundation\Response;
  *           Symfony2:
  *              app_path: 'app/front'
  *              environment: 'local_test'
+ *              debug: true
  *
  * ## Public Properties
  *
@@ -52,7 +54,7 @@ class Symfony2 extends \Codeception\Lib\Framework
      */
     public $container;
 
-    public $config = array('app_path' => 'app', 'environment' => 'test');
+    public $config = array('app_path' => 'app', 'environment' => 'test', 'debug' => true);
     
     /**
      * @var
@@ -67,7 +69,7 @@ class Symfony2 extends \Codeception\Lib\Framework
         if (!file_exists($cache)) throw new ModuleRequire(__CLASS__, 'Symfony2 bootstrap file not found in '.$cache);
         require_once $cache;
         $this->kernelClass = $this->getKernelClass();
-        $this->kernel = new $this->kernelClass($this->config['environment'], true);
+        $this->kernel = new $this->kernelClass($this->config['environment'], $this->config['debug']);
     }
     
     public function _before(\Codeception\TestCase $test) {