+++ src/Codeception/TestCase/Interfaces/ScenarioDriven.php
@@ -10,7 +10,7 @@ interface ScenarioDriven {
      */
     public function getScenario();
 
-    public function getScenarioText();
+    public function getScenarioText($format = 'text');
 
     public function preload();
 }