+++ tests/data/app/index.php
@@ -22,8 +22,9 @@ $urls = array(
     '/redirect' => 'redirect',
     '/redirect2' => 'redirect2',
     '/redirect3' => 'redirect3',
+    '/redirect4' => 'redirect4',
     '/facebook\??.*' => 'facebookController',
-    '/form/(field|select|checkbox|file|textarea|hidden|complex|button|radio|select_multiple|unchecked|empty|popup|example1|example2|example3)(#|\?.*?)?' => 'form',
+    '/form/(.*?)(#|\?.*?)?' => 'form',
     '/articles\??.*' => 'articles'
 );