+++ CHANGELOG.md
@@ -5,6 +5,40 @@
 * Conflicts API added to modules
 
 
+#### 2.0.1 06/21/2014
+
+* [Phalcon1] Fixed connector
+* [WebDriver] added seeInPageSource and dontSeeInPageSource methods
+* [WebDriver] see method now checks only for visible BODY element by @artyfarty
+* [REST] added Bearer authentication by @dizews
+* removed auto added submit buttons in forms previously used as hook for DomCrawler
+
+
+#### 2.0.1 06/20/2014
+
+* BUGFIX: PHP 5.4.x compatibility fixed. Sample error output: 'Method WelcomeCept.php does not exist' #1084 #1069 #1109
+
+
+#### 2.0.1 06/19/2014
+
+* Second parameter of Cest method is treated as scenario variable on parse. Fix #1058
+
+
+#### 2.0.1 06/18/2014
+
+* prints raw stack trace including codeception classes in -vvv mode
+* screenshots on fail are saved to properly named files #1075
+* [Symfony2] added debug config option to switch debug mode by @pmcjury
+
+
+
+#### 2.0.0 06/05/2014
+
+* renamed `_logs` dir to `_output` by default
+* renamed `_helpers` dir to `_support` by default
+* Guy renamed to Tester
+* Bootstrap command got 3 installation modes: default, compat, setup
+
 #### 2.0.0 05/27/2014
 
 * added --coverage-text option